format run command test case
