+++ src/azure-cli/azure/cli/command_modules/acs/tests/latest/mocks.py
@@ -56,3 +56,12 @@ class MockCmd:
 
     def get_models(self, *attr_args, **kwargs):
         return self.cmd.get_models(*attr_args, **kwargs)
+
+
+class MockUrlretrieveUrlValidator(object):
+    def __init__(self, url, version):
+        self.url = url
+        self.version = version
+
+    def __eq__(self, other):
+        return other.startswith(self.url) and self.version in other
